ringbuf looking better
